changed age type int to byte

+++ src/main/java/web/config/DBinit.java
@@ -23,7 +23,7 @@ public class DBinit {
         User user1 = new User();
         user1.setFirstName("Tom");
         user1.setLastName("Jones");
-        user1.setAge(27);
+        user1.setAge((byte) 27);
         user1.setEmail("tom@gmail.com");
         user1.setPassword("tom");
         user1.addRole(roleService.getRoleByName("ADMIN"));
@@ -31,7 +31,7 @@ public class DBinit {
         User user2 = new User();
         user2.setFirstName("Анна");
         user2.setLastName("Волкова");
-        user2.setAge(24);
+        user2.setAge((byte) 24);
         user2.setEmail("ann@mail.ru");
         user2.setPassword("анна");
         user2.addRole(roleService.getRoleByName("USER"));
@@ -39,7 +39,7 @@ public class DBinit {
         User user3 = new User();
         user3.setFirstName("Sam");
         user3.setLastName("Black");
-        user3.setAge(35);
+        user3.setAge((byte) 35);
         user3.setEmail("sam@yahoo.com");
         user3.setPassword("sam");
         user3.addRole(roleService.getRoleByName("ADMIN"));
